fluent-syntax: Avoid forEach to make code more minification-friendly
This fixes an obscure bug in uglify-es. The minified code inlined the
withSpans decorator in a way which put the fn argument in an outer
scope.  The decorated function closed over fn and when the time came to
call it, it would call a different function than the one that was
decorated.

Using a regular for-of loop seems to have fixed the bug. Tested using:

    "uglify-es": "^3.3.4",
    "uglifyjs-webpack-plugin": "1.2.5",
    "react-scripts": "^2.0.0-next.3e165448"