Add node 10 and 12 to test matrix, disable pl pluralrules test.
Also, I'm making node 8 not specify the subrevision, there are
newer versions of it by now, and we're not specific on 10 nor 12.

The test for Polish plural forms requires that Intl.PluralRules
exists for Polish, and the standard node distributions don't
ship with ICU. Let's rely on the upstream tests.