Default the child element of Localized to null
If no child element is provided and <Localized /> tries to return
the children as a fallback, undefined is returned from the
component. This causes a React error, which is difficult to
track down. Instead, default the child element to null, which is
a valid return value for a React component.