Return ReactElement from Localized
React typings (wrongly) expect the return type of functional components
to be a ReactElement, thus making () => "Hello" invalid. This is
incorrect; as of React 16, it's possible to return plain strings from
functional components and Component.render(). See DefinitelyTyped#18912
and TypeScript#21699.

To work around this, annotate the return value of Localized to be of
type ReactElement, and whenever a string or null are returned, wrap them
in Fragments.