Migrate @fluent/react to TypeScript (#458)
* Migrate @fluent/react to TypeScript

* Fix paths in fluent-gecko

* Return ReactElement from Localized

React typings (wrongly) expect the return type of functional components
to be a ReactElement, thus making () => "Hello" invalid. This is
incorrect; as of React 16, it's possible to return plain strings from
functional components and Component.render(). See DefinitelyTyped#18912
and TypeScript#21699.

To work around this, annotate the return value of Localized to be of
type ReactElement, and whenever a string or null are returned, wrap them
in Fragments.

* Add the build make target

* Fix compat.js make target

* Export props interfaces

* Export MarkupParser type

* Fix return type of withLocalization

* Convert elems keys to lowercase; lookup via Element.localName

* Use Node.nodeName.toLowerCase() for elems lookup

* Add a comment about bundle.getMessage(id)!

* Move src/vendor out of src/