Spy on console.warn in tests, and throw an error if called
The current tests aren't explicit on testing the behavior of what
errors get reported. This makes it difficult to understand the
current behavior. This commit adds a mechanism to fail a test if
the console.warn is called unexpectedly. Following commits will fix
up tests that fail from this new mechanism.