Run prettier on test files to minimize inline snapshot code diff
Inline snapshots in Jest are powered by Prettier, however prettier
does not appear to be hooked up to the eslint.

https://jestjs.io/docs/en/snapshot-testing#inline-snapshots

In the following commits I add new inline snapshot tests, which
invalidates a lot of the existing formatting. This commit runs prettier
on the entire test directory to ensure that the following commits have
clean code diffs.