Throw an error when the LocalizationProvider is not properly initialized
The current behavior is permissive, and works without warnings. This
change makes it so that omitting the LocalizationProvider throws an
error. I was originally trying to decide between throwing an error and
emitting a console error. I opted for an error to have the behavior in
line with what the react-redux Provider does.

This is potentially a breaking change for users that are relying on the
LocalizationProvider being omitted before the bundles are loaded. This
behavior can still be achieved by providing an empty bundle list to the
LocalizationProvider.