Set type: "module" for esm/ and test/ in all packages & update to mocha@9 (#567)
* feat: Add an esm/package.json to all with { "type": "module" }

While the top-level package.json files include a "type": "commonjs"
specification, the esm/ directories can (and ought) to either have
their contents use the .mjs extension, or include their own
package.json file with the correct "type".

This adds the package.json while keeping the extensions as .js, as
this is significantly easier for the TypeScript interop; proper .mjs
support is currently an unresolved issue:
https://github.com/microsoft/TypeScript/issues/18442

* refactor: Use .js extension for all ES imports

While extensionless imports are supported by TypeScript as well as
the esm package, the extensions are required by Node.js when
resolving imports. TypeScript's tsc is also completely fine with
using .js extensions when referring to .ts files.

* test(fluent-dom): Move elem() from index.js to util.js

* test: Update to mocha@9, now using Node.js native import

* chore: Patch ESLint configs to allow linting to succeed

The consideration of test/ files as ES modules requires a few of them
to use import.meta, which isn't supported by the default parser of
eslint@6. So this is a minimal patch to get linting to continue
working until ESLint itself is updated.